+++ src/gui/gui-bar-item.c
@@ -435,7 +435,7 @@ gui_bar_item_get_value (struct t_gui_bar *bar, struct t_gui_window *window,
                 time_diff = util_timeval_diff (&start_time, &end_time);
                 if (time_diff >= debug_long_callbacks)
                 {
-                    str_diff = util_get_microseconds_string (time_diff);
+                    str_diff = util_get_microseconds_string ((unsigned long long)time_diff);
                     log_printf (
                         _("debug: long callback: bar: %s, item: %s, plugin: %s, "
                           "time elapsed: %s"),
